requirements.txt
pandas==2.3.1
numpy==2.0.0
scikit-learn==1.5.0
catboost==1.2.8
fastapi==0.116.1
cloudpickle==3.1.1
pyarrow==16.1.0
prefect==2.20.10
evidently==0.7.10
httpx
mlflow==2.22.1
uvicorn
pydantic==1.10.22
pytest
google-cloud-storage

#$env:PIPENV_TIMEOUT=60
#pipenv install -r requirements.txt
